# src/main/resources/openapi/registre-api.yaml
openapi: 3.0.3
info:
  title: Metadata-Collecte API
  version: 0.0.1
  description: |
    ## Codes-List Registry
    - Registry of codes-list
    - Registry of suggester configuration
    ## Questionnaire Registry
    - Registry of questionnaire
    - Registry of questionnaires projection and its variables
servers:
  - url: http://localhost:8080
...
paths:
...
  /codes-lists:
    post:
      tags:
        - Codes List Publication
      summary: Create codes list (metadata only)
      operationId: createCodesList
      description: |
        Admin only. Create a code list without content or search config.
        These must be added later via PUT endpoints.
      requestBody:
        $ref: "#/components/requestBodies/CodesListDto"
      responses:
        "201":
          description: Created
        "409":
          $ref: "#/components/responses/ErrorResponse"

    get:
      tags:
        - Codes List Recovery
      summary: Get all codes lists metadata
      operationId: getAllCodesLists
      responses:
        "200":
          $ref: "#/components/responses/CodesListsMetadata"
...
  /codes-lists/{codesListId}/search-configuration:
    put:
      tags:
        - Codes List Publication
      summary: Add search configuration to a codes list
      operationId: putCodesListSearchConfigById
      parameters:
        - name: codesListId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        $ref: "#/components/requestBodies/CodesListSearchConfiguration"
      responses:
        "201":
          description: Search configuration set successfully
        "409":
          $ref: "#/components/responses/ErrorResponse"

    get:
      tags:
        - Codes List Recovery
      summary: Get search configuration
      operationId: getCodesListSearchConfigById
      parameters:
        - name: codesListId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/CodesListSearchConfiguration"
  /codes-lists/{codesListId}:
    get:
      tags:
        - Codes List Recovery
      summary: Get full content of a codes list
      operationId: getCodesListById
      parameters:
        - name: codesListId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/CodesList"
  /codes-lists/{codesListId}/metadata:
    get:
      tags:
        - Codes List Recovery
      summary: Get codes list metadata
      operationId: getCodesListMetadataById
      parameters:
        - name: codesListId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/CodesListMetadata"
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 409
        message:
          type: string
          example: Content of ${codeListId} already exists
...
    CodesListDto:
      type: object
      properties:
        metadata:
          $ref: "#/components/schemas/Metadata"
        searchConfiguration:
          $ref: "#/components/schemas/CodesListSearchConfiguration"
        content:
          $ref: "#/components/schemas/CodesList"
    Metadata:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: UUID created during creation by api
        label:
          type: string
        version:
          type: string
        externalLink:
          $ref: "#/components/schemas/CodesListExternalLink"
    CodesListSearchConfiguration:
      type: object
    CodesListExternalLink:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          description: UUID from RMeS
        version:
          type: string
          description: version of code-list from RMeS
    CodesList:
      type: array
      items:
        $ref: "#/components/schemas/Code"
    Code:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
      additionalProperties: true
  requestBodies:
...
    CodesListDto:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CodesListDto"
...
    CodesListSearchConfiguration:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CodesListSearchConfiguration"
...
  responses:
...
    CodesList:
      description: Codes list content
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CodesList"
    CodesListsMetadata:
      description: List of codes list metadata
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/Metadata"
    CodesListMetadata:
      description: Metadata of a codes list
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Metadata"
    CodesListSearchConfiguration:
      description: Search configuration
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CodesListSearchConfiguration"
    ErrorResponse:
      description: Structured error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"